Update security rule with public IP address
On starting the app, the user's public IP address is queried using
ident.me. If the public IP address cannot be determined (e.g. if the
user if offline) the process exits with an error message.

The promptForSecurityRule message now returns the entire rule instead of
just the rule ID. This is because the SDK method to update the security
rule needs all of the properties to be set. All properties other than
the IP address are read from the existing rule and reapplied.